Update layout of existing destination panel + add some improvements
