+++ apps/studio/components/interfaces/Database/Replication/DestinationPanel/DestinationPanel.tsx
@@ -7,6 +7,7 @@ import { toast } from 'sonner'
 import * as z from 'zod'
 
 import { useFlag, useParams } from 'common'
+import { CreateAnalyticsBucketSheet } from 'components/interfaces/Storage/AnalyticsBuckets/CreateAnalyticsBucketSheet'
 import { getKeys, useAPIKeysQuery } from 'data/api-keys/api-keys-query'
 import { useProjectSettingsV2Query } from 'data/config/project-settings-v2-query'
 import { useCheckPrimaryKeysExists } from 'data/database/primary-keys-exists-query'
@@ -91,6 +92,7 @@ export const DestinationPanel = ({
   const editMode = !!existingDestination
   const [showDisclaimerDialog, setShowDisclaimerDialog] = useState(false)
   const [publicationPanelVisible, setPublicationPanelVisible] = useState(false)
+  const [newBucketSheetVisible, setNewBucketSheetVisible] = useState(false)
   const [pendingFormValues, setPendingFormValues] = useState<z.infer<typeof FormSchema> | null>(
     null
   )
@@ -438,9 +440,9 @@ export const DestinationPanel = ({
     <>
       <Sheet open={visible} onOpenChange={onClose}>
         <SheetContent
+          size="lg"
           showClose={false}
-          size="default"
-          className={publicationPanelVisible ? 'right-32' : 'right-0'}
+          className={publicationPanelVisible || newBucketSheetVisible ? 'right-20' : 'right-0'}
         >
           <div className="flex flex-col h-full" tabIndex={-1}>
             <SheetHeader>
@@ -458,13 +460,12 @@ export const DestinationPanel = ({
               ) : (
                 <Form_Shadcn_ {...form}>
                   <form id={formId} onSubmit={form.handleSubmit(onSubmit)}>
-                    <div className="p-5">
-                      <p className="text-sm font-medium text-foreground mb-1">
-                        Destination details
-                      </p>
-                      <p className="text-sm text-foreground-light mb-4">
-                        Name your destination and choose which data to replicate
-                      </p>
+                    <DestinationTypeSelection form={form} editMode={editMode} />
+
+                    <DialogSectionSeparator />
+
+                    <div className="p-5 flex flex-col gap-y-6">
+                      <p className="text-sm font-medium text-foreground">Destination details</p>
 
                       <div className="space-y-4">
                         <DestinationNameInput form={form} />
@@ -476,23 +477,21 @@ export const DestinationPanel = ({
                         />
                       </div>
                     </div>
+
                     <DialogSectionSeparator />
-                    <DestinationTypeSelection form={form} editMode={editMode} />
+
                     {selectedType === 'BigQuery' && etlEnableBigQuery ? (
-                      <>
-                        <DialogSectionSeparator />
-                        <BigQueryFields form={form} />
-                      </>
+                      <BigQueryFields form={form} />
                     ) : selectedType === 'Analytics Bucket' && etlEnableIceberg ? (
-                      <>
-                        <DialogSectionSeparator />
-                        <AnalyticsBucketFields
-                          form={form}
-                          setIsFormInteracting={setIsFormInteracting}
-                        />
-                      </>
+                      <AnalyticsBucketFields
+                        form={form}
+                        setIsFormInteracting={setIsFormInteracting}
+                        onSelectNewBucket={() => setNewBucketSheetVisible(true)}
+                      />
                     ) : null}
+
                     <DialogSectionSeparator />
+
                     <AdvancedSettings form={form} />
                   </form>
                 </Form_Shadcn_>
@@ -524,6 +523,11 @@ export const DestinationPanel = ({
         sourceId={sourceId}
         visible={publicationPanelVisible}
         onClose={() => setPublicationPanelVisible(false)}
+      />
+
+      <CreateAnalyticsBucketSheet
+        open={newBucketSheetVisible}
+        onOpenChange={setNewBucketSheetVisible}
       />
 
       <ReplicationDisclaimerDialog
